+++ prometheus_ecs_discoverer/run.py
@@ -7,8 +7,7 @@ from botocore.config import Config
 from loguru import logger
 from prometheus_client import start_http_server
 
-from prometheus_ecs_discoverer import (discovery, fetching, marshalling, s,
-                                       telemetry)
+from prometheus_ecs_discoverer import discovery, fetching, marshalling, s, telemetry
 
 # Copyright 2020 Tim Schwenke. Licensed under the Apache License 2.0